Legacy/requirements.txt
Flask
Pillow
numpy
pyyaml
gunicorn
flask-caching

